+++ CHANGELOG.md
@@ -11,6 +11,9 @@ to a section with the version name.
 * Add support for precomputed invoices from POS devices
 * Ignore permissions when inserting ZATCA integration log
 * Fix item total amount in sales invoice additional fields for return invoices
+* Abort submission for sales invoice additional field document if the status is resend.
+* Update ZATCA integration log with Zatca status.
+* Autoname method for ZATCA integration log.
 
 ## 0.5.0
 
@@ -20,6 +23,7 @@ to a section with the version name.
   * Sort additional fields by creation (oldest first)
   * Run in batches (of 100 by default)
   * Add more logging
+
 
 ## 0.4.0