+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to a section with the version name.
 ## Unreleased Changes
 
 * Add support for precomputed invoices from POS devices
+* Make precomputed invoice and sales invoice additional fields UUID unique to safeguard against bugs causing double ZATCA submissions
 
 ## 0.5.0